LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/base-style-invalidation-expected.txt


FAIL The chosen position options changes when the base style differs assert_equals: expected 20 but got 0
